Move util classes to reasonable packages
