+++ core/esmf-util/src/main/java/org/eclipse/esmf/util/download/ProxyConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2026 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package org.eclipse.esmf.aspectmodel.resolver;
+package org.eclipse.esmf.util.download;
 
 import java.net.Authenticator;
 import java.net.InetSocketAddress;
